[BREAKINGCHANGE] TracingGanttChart: configure attribute links in panel spec instead of panel props (#210)
* [BREAKINGCHANGE] TracingGanttChart: configure attribute links in panel spec instead of panel props

Previously, the attribute links (linking from span attributes to custom
pages) were configured via panel props. However, this works only if the
panel is embedded directly, i.e. `<TracingGanttChart.PanelComponent .../>`
and doesn't work when included via `<Panel definition={{...}} />`.

This was a problem, because the TempoExplorer in the Tempo plugin
includes the panel via the plugin system (via `<Panel ... />`) and
therefore cannot pass custom props to the panel plugin.

This change moves the attribute link configuration to the panel spec.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* don't duplicate RouterComponent type

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* use RouterContext

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* make RouterComponent optional

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* fix types

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

---------

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>