[ENHANCEMENT] TracingGanttChart: support linking to Gantt Chart with a span selected (#282)
* [ENHANCEMENT] TracingGanttChart: support linking to Gantt Chart with a span selected

This change adds a `selectedSpanId` property to the Gantt Chart. When
this property is set, the referenced span is visible (i.e. on top of the
current viewport) and selected (highlighted and the detail pane is open).

This is mainly used to highlight the referenced span when clicking on a span link.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* fix test

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

---------

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>